+++ src/apis/auth.ts
@@ -2,11 +2,19 @@ const fetch = require('node-fetch');
 
 import { AuthReturnDataforRestaurantManagementGroupClient } from '../types/auth.types';
 import { ToastError } from '..';
+import { URLSearchParams } from 'url';
+
 
 export async function requestNewToken(toastHostname: string, toastClient: string, toastSecret: string): Promise<AuthReturnDataforRestaurantManagementGroupClient> {
+    
+    const params = new URLSearchParams();
+    params.append('grant_type', 'client_credentials');
+    params.append('client_id', toastClient);
+    params.append('client_secret', toastSecret);
+    
     const options = {
         method: 'POST',
-        body: `grant_type=client_credentials&client_id=${toastClient}&client_secret=${toastSecret}`,
+        body: params,
         headers: {
             'Content-Type': 'application/x-www-form-urlencoded'
         },